fastapi_backend/requirements.txt
fastapi
uvicorn
diffusers
transformers
torch
peft
pillow
huggingface_hub
gunicorn